Move file fixes to the very end of verify tasks.
To ensure we're leaving platforms and sites in a usable state,
we need to ensure that we run our scripts last, to account for
other operations during verify (such as CiviCRM's cache-clears.

We do this by prepending a zz_ prefix to our Drush commandfiles
and hook implementations, since these are called in alphabetical
order.